Add files via upload
For each integer  n in the inclusive interval a,b:

If a<=n<=b , then print the English representation of it in lowercase. That is "one" ,"two"  and so on.
Else if  b<=n,it is an even number, then print "even" and it is an odd number, then print "odd".